rest2

+++ projet/serviceRest2/src/main/java/com/rest2/service/ParticipationService.java
@@ -2,6 +2,7 @@ package com.rest2.service;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Map;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
@@ -15,6 +16,9 @@ import com.rest2.repository.ParticipationRepository;
 
 @Service
 public class ParticipationService {
+
+    private static final String REST1_URL = System.getenv().getOrDefault("REST1_SERVICE_URL",
+            "http://localhost:8081/rest1/");
     private final ParticipationRepository participationRepository;
     private final RestTemplate restTemplate;
 
@@ -25,15 +29,20 @@ public class ParticipationService {
 
     public boolean participer(int idUtilisateur, int idCompetition) {
 
-        String url = UriComponentsBuilder.fromUriString("http://localhost:8081/rest1/getCompetitions")
-                                        .queryParam("idCompetition", idCompetition)
-                                        .toUriString();
+        String url1 = UriComponentsBuilder.fromUriString(REST1_URL + "getCompetitions")
+                .queryParam("idCompetition", idCompetition)
+                .toUriString();
+        String url2 = UriComponentsBuilder.fromUriString(REST1_URL + "getUtilisateurs")
+                .queryParam("id", idUtilisateur)
+                .toUriString();
 
-        ResponseEntity<String> response = restTemplate.getForEntity(url,String.class);
+        ResponseEntity<Map> response1 = restTemplate.getForEntity(url1, Map.class);
+        ResponseEntity<Map> response2 = restTemplate.getForEntity(url2, Map.class);
 
-        if (response.getStatusCode().is2xxSuccessful()) {
-            String responseBody = response.getBody();
-            if (responseBody != null) {
+        if (response1.getStatusCode().is2xxSuccessful() && response2.getStatusCode().is2xxSuccessful()) {
+            Map responseBody1 = response1.getBody();
+            Map responseBody2 = response2.getBody();
+            if (responseBody1.get("data") != null && responseBody2.get("data") != null) {
 
                 ParticipationId newParticipationId = new ParticipationId();
                 newParticipationId.setPfkUtilisateur(idUtilisateur);
@@ -46,7 +55,7 @@ public class ParticipationService {
                 return true;
             } else {
                 return false; // Si l'API renvoie une erreur ou un statut différent
-        
+
             }
         } else {
             return false; // Si l'API retourne une erreur (code autre que 2xx)
@@ -54,25 +63,48 @@ public class ParticipationService {
     }
 
     public boolean desinscrire(int idUtilisateur, int idCompetition) {
-        ParticipationId participationId = new ParticipationId();
-        participationId.setPfkUtilisateur(idUtilisateur);
-        participationId.setPfkCompetition(idCompetition);
+        String url1 = UriComponentsBuilder.fromUriString(REST1_URL + "getCompetitions")
+                .queryParam("idCompetition", idCompetition)
+                .toUriString();
+        String url2 = UriComponentsBuilder.fromUriString(REST1_URL + "getUtilisateurs")
+                .queryParam("id", idUtilisateur)
+                .toUriString();
 
-        // Vérifier si l'entité existe avant de la supprimer
-        if (participationRepository.existsById(participationId)) {
-            participationRepository.deleteById(participationId);
-            return true;
+        ResponseEntity<Map> response1 = restTemplate.getForEntity(url1, Map.class);
+        ResponseEntity<Map> response2 = restTemplate.getForEntity(url2, Map.class);
+
+        if (response1.getStatusCode().is2xxSuccessful() && response2.getStatusCode().is2xxSuccessful()) {
+            Map responseBody1 = response1.getBody();
+            Map responseBody2 = response2.getBody();
+            if (responseBody1.get("data") != null && responseBody2.get("data") != null) {
+                ParticipationId participationId = new ParticipationId();
+                participationId.setPfkUtilisateur(idUtilisateur);
+                participationId.setPfkCompetition(idCompetition);
+
+                // Vérifier si l'entité existe avant de la supprimer
+                if (participationRepository.existsById(participationId)) {
+                    participationRepository.deleteById(participationId);
+                    return true;
+                } else {
+                    return false;
+                }
+            } else {
+                return false; // Si l'API renvoie une erreur ou un statut différent
+
+            }
         } else {
-            return false;
+            return false; // Si l'API retourne une erreur (code autre que 2xx)
         }
     }
 
     public List<ParticipationDTO> getParticipations(int idCompetition) {
+        
         List<Participation> participations = participationRepository.findById_PfkCompetition(idCompetition);
         List<ParticipationDTO> participationDTOs = new ArrayList<>();
 
         for (Participation participation : participations) {
-            ParticipationDTO participationDTO = new ParticipationDTO(participation.getId().getPfkCompetition(), participation.getId().getPfkUtilisateur());
+            ParticipationDTO participationDTO = new ParticipationDTO(participation.getId().getPfkCompetition(),
+                    participation.getId().getPfkUtilisateur());
             participationDTOs.add(participationDTO);
         }
         return participationDTOs;
